requirements.txt
#required
PyYAML>=5.3
epyestim
matplotlib==3.1.3
more-itertools==8.2.0
numpy==1.18.1
pandas==1.0.1
seaborn==0.10.0

#optional for save yaml loading and loading environment variables
yamlordereddictloader>0.4
python-dotenv>=0.12
